Part A
In part A I implemented a discrete BPSK with OVERSAMPLE = 1. First, I generate a vector of 0, 1 with equal probability.  Then this input vector is modulated 0 -> -1, 1 -> 1.  Then I generate the symbols for this vector. Since it is BPSK, the symbols will have  its Im = 0. Then for each value of SNR (in dB), I calculate the output of the channel using the Matlab's function awgn, which adds to the input signal an additive white gaussian noise. The next step is to decide whether the output bit is 0 or 1 and after that check if it has error. At the end I calculate the BER and finally graph the result BER vs BER-theory